requirements.txt
flask
tensorflow==2.8.0
numpy
pillow
protobuf==3.20.*
